+++ src/source/periodic.rs
@@ -1,8 +1,10 @@
 use std::time::Duration;
 
 use super::SeekError;
-use crate::common::{ChannelCount, SampleRate};
-use crate::Source;
+use crate::{
+    common::{ChannelCount, SampleRate},
+    Source,
+};
 
 /// Internal function that builds a `PeriodicAccess` object.
 pub fn periodic<I, F>(source: I, period: Duration, modifier: F) -> PeriodicAccess<I, F>
@@ -10,20 +12,15 @@ where
     I: Source,
 {
     // TODO: handle the fact that the samples rate can change
-    // TODO: generally, just wrong
-    let update_ms = period.as_secs() as u32 * 1_000 + period.subsec_millis();
-    let update_frequency =
-        (update_ms * source.sample_rate().get()) / 1000 * source.channels().get() as u32;
+    let update_frequency = (period.as_secs_f32()
+        * (source.sample_rate().get() as f32)
+        * (source.channels().get() as f32)) as u32;
 
     PeriodicAccess {
         input: source,
         modifier,
         // Can overflow when subtracting if this is 0
-        update_frequency: if update_frequency == 0 {
-            1
-        } else {
-            update_frequency
-        },
+        update_frequency: update_frequency.max(1),
         samples_until_update: 1,
     }
 }