+++ CHANGELOG.md
@@ -9,8 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- `Chirp` now implements `Iterator::size_hint` and `ExactSizeIterator`.
+
 ### Fixed
 - docs.rs will now document all features, including those that are optional.
+- `Chirp::next` now returns `None` when the total duration has been reached, and will work
+  correctly for a number of samples greater than 2^24.
+- `PeriodicAccess` is slightly more accurate for 44.1 kHz sample rate families.
 
 ## Version [0.21.1] (2025-07-14)